requirements.txt
asgiref==3.2.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
codecov==2.0.16
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
coveralls==1.11.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-model-utils==4.0.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
drf-yasg==1.17.1
factory-boy==2.12.0
Faker==0.9.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mixer==6.1.3
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.1
pluggy==0.13.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
pytest-assert-utils==0.1.0
pytest-common-subject==1.0.1
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-drf==0.1.0
pytest-fixture-order==0.1.1
pytest-lambda==1.1.0
python-coveralls==2.9.3
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
whitenoise==5.0.1
wrapt==1.12.0
zipp==3.0.0
